In chlDelUserRE, also delete r_user_group rows where the group_user_id matches the user or group being deleted.  Previously, when groups were being deleted, if there were users that were part of the group, a row per such user would be left in the table.  This wouldn't really hurt anything, except make the table grow on occasion.
git-svn-id: svn://irodssvn.ucsd.edu/trunk@2396 ea137bd6-aa86-47b6-b90d-b101d9f19a90

+++ iRODS/server/icat/src/icatHighLevelRoutines.c
@@ -1698,11 +1698,12 @@ int chlDelUserRE(rsComm_t *rsComm, userInfo_t *userInfo) {
    }
 
    /* Remove both the special user_id = group_user_id entry and any
-      other access entries for this user */
+      other access entries for this user (or group) */
+   cllBindVars[cllBindVarCount++]=iValStr;
    cllBindVars[cllBindVarCount++]=iValStr;
    if (logSQL) rodsLog(LOG_SQL, "chlDelUserRE SQL 4");
    status = cmlExecuteNoAnswerSql(
-	    "delete from r_user_group where user_id=?",
+	    "delete from r_user_group where user_id=? or group_user_id=?",
 	    &icss);
    if (status!=0 && status != CAT_SUCCESS_BUT_WITH_NO_INFO) {
       int i;
